test(label): migrate basic
