+++ core/src/components/label/test/basic/label.e2e.ts-snapshots/label-basic-ios-rtl-Mobile-Chrome-linux.png
