+++ core/src/components/label/test/basic/label.e2e.ts-snapshots/label-basic-md-ltr-Mobile-Safari-linux.png
