+++ core/src/components/label/test/basic/label.e2e.ts-snapshots/label-fixed-ios-ltr-Mobile-Firefox-linux.png
