+++ core/src/components/label/test/basic/label.e2e.ts-snapshots/label-fixed-md-ltr-Mobile-Chrome-linux.png
