+++ core/src/components/label/test/basic/label.e2e.ts-snapshots/label-fixed-md-rtl-Mobile-Chrome-linux.png
